Update README.md
“Online Marriage Registration and Information System” is a web application. Our main goal
is to make a platform for both marriage registration and finding a bride or groom. The existing
marriage registration system has so many failings and needs to be updated. We are trying
to take both registrations and finding a bride, groom on a platform. We will add some exciting
features like live chats and recommendations for our project. We hope that, after developing
this website, it will bring a remarkable change to the existing feature of our system.